# Conflicts:
#	dubbo-dependencies-bom/pom.xml
#	dubbo-dependencies/dubbo-dependencies-zookeeper/pom.xml
#	pom.xml

+++ dubbo-config/dubbo-config-api/src/main/java/org/apache/dubbo/config/bootstrap/DubboBootstrap.java
@@ -67,7 +67,6 @@ import org.apache.dubbo.registry.client.ServiceDiscovery;
 import org.apache.dubbo.registry.client.ServiceDiscoveryRegistry;
 import org.apache.dubbo.registry.client.ServiceInstance;
 import org.apache.dubbo.registry.support.AbstractRegistryFactory;
-import org.apache.dubbo.rpc.Protocol;
 import org.apache.dubbo.rpc.model.ApplicationModel;
 
 import java.io.IOException;
@@ -134,6 +133,8 @@ public class DubboBootstrap extends GenericEventListener {
     private final Lock lock = new ReentrantLock();
 
     private final Condition condition = lock.newCondition();
+
+    private final Lock destroyLock = new ReentrantLock();
 
     private final ExecutorService executorService = newSingleThreadExecutor();
 
@@ -1033,44 +1034,29 @@ public class DubboBootstrap extends GenericEventListener {
     }
 
     public void destroy() {
-        // for compatibility purpose
-        DubboShutdownHook.destroyAll();
-
-        if (started.compareAndSet(true, false)
-                && destroyed.compareAndSet(false, true)) {
-
-            unregisterServiceInstance();
-            unexportMetadataService();
-            unexportServices();
-            unreferServices();
-
-            destroyRegistries();
-            destroyProtocols();
-            destroyServiceDiscoveries();
-
-            clear();
-            shutdown();
-            release();
-        }
-    }
-
-    /**
-     * Destroy all the protocols.
-     */
-    private void destroyProtocols() {
-        ExtensionLoader<Protocol> loader = ExtensionLoader.getExtensionLoader(Protocol.class);
-        for (String protocolName : loader.getLoadedExtensions()) {
+        if (destroyLock.tryLock()) {
             try {
-                Protocol protocol = loader.getLoadedExtension(protocolName);
-                if (protocol != null) {
-                    protocol.destroy();
+                DubboShutdownHook.destroyAll();
+
+                if (started.compareAndSet(true, false)
+                        && destroyed.compareAndSet(false, true)) {
+
+                    unregisterServiceInstance();
+                    unexportMetadataService();
+                    unexportServices();
+                    unreferServices();
+
+                    destroyRegistries();
+                    DubboShutdownHook.destroyProtocols();
+                    destroyServiceDiscoveries();
+
+                    clear();
+                    shutdown();
+                    release();
                 }
-            } catch (Throwable t) {
-                logger.warn(t.getMessage(), t);
+            } finally {
+                destroyLock.unlock();
             }
-        }
-        if (logger.isDebugEnabled()) {
-            logger.debug(NAME + "'s all ProtocolConfigs have been destroyed.");
         }
     }
